backend/requirements.txt
Flask==2.3.2
Flask-Cors==5.0.1
flask-mysql-connector==1.1.0
mysql-connector-python==9.3.0
python-dotenv==0.21.0
PyJWT==2.6.0
python-dateutil==2.8.2
PyMySQL==1.1.0 